Handle factory image or avb key not provided

+++ flasher.go
@@ -91,7 +91,15 @@ func main() {
 		}
 	}
 	checkPrerequisiteFiles()
-	if factoryImage == "" {
+	//TODO see if there's a better way of getting a list of google devices
+	googleDevice := false
+	for _, codename := range []string{"sailfish", "marlin", "walleye", "taimen", "blueline", "crosshatch", "sargo", "bonito"} {
+		if codename == device {
+			googleDevice = true
+			break
+		}
+	}
+	if factoryImage == "" && googleDevice {
 		fmt.Println("Factory image missing. Attempting to download from https://developers.google.com/android/images/index.html")
 		err = getFactoryImage()
 		if err != nil {
@@ -100,28 +108,30 @@ func main() {
 			os.Exit(1)
 		}
 	}
-	err = extractZip(path.Base(factoryImage), cwd)
-	if err != nil {
-		fmt.Println(err.Error())
-		fmt.Println("Cannot continue without the device factory image. Exiting...")
-		os.Exit(1)
-	}
-	factoryImage = regexp.MustCompile(".*\\.[0-9]{3}").FindAllString(factoryImage, -1)[0]
-	factoryImage = cwd + string(os.PathSeparator) + factoryImage + string(os.PathSeparator)
-	files, err := ioutil.ReadDir(factoryImage)
-	if err != nil {
-		fmt.Println(err.Error())
-		fmt.Println("Cannot continue without the device factory image. Exiting...")
-		os.Exit(1)
-	}
-	for _, file := range files {
-		file := file.Name()
-		if strings.Contains(file, "bootloader") {
-			bootloader = factoryImage + file
-		} else if strings.Contains(file, "radio") {
-			radio = factoryImage + file
-		} else if strings.Contains(file, "image") {
-			image = factoryImage + file
+	if factoryImage != "" {
+		err = extractZip(path.Base(factoryImage), cwd)
+		if err != nil {
+			fmt.Println(err.Error())
+			fmt.Println("Cannot continue without the device factory image. Exiting...")
+			os.Exit(1)
+		}
+		factoryImage = regexp.MustCompile(".*\\.[0-9]{3}").FindAllString(factoryImage, -1)[0]
+		factoryImage = cwd + string(os.PathSeparator) + factoryImage + string(os.PathSeparator)
+		files, err := ioutil.ReadDir(factoryImage)
+		if err != nil {
+			fmt.Println(err.Error())
+			fmt.Println("Cannot continue without the device factory image. Exiting...")
+			os.Exit(1)
+		}
+		for _, file := range files {
+			file := file.Name()
+			if strings.Contains(file, "bootloader") {
+				bootloader = factoryImage + file
+			} else if strings.Contains(file, "radio") {
+				radio = factoryImage + file
+			} else if strings.Contains(file, "image") {
+				image = factoryImage + file
+			}
 		}
 	}
 	flashDevices()
@@ -147,9 +157,6 @@ func checkPrerequisiteFiles() {
 	}
 	if altosImage == "" {
 		fmt.Println("Cannot continue without altOS device image. Exiting...")
-		os.Exit(1)
-	} else if altosKey == "" {
-		fmt.Println("Cannot continue without altOS key. Exiting...")
 		os.Exit(1)
 	}
 }
@@ -263,6 +270,7 @@ func getFactoryImage() error {
 		return err
 	}
 	body := string(out)
+	//TODO better pattern matching for links. maybe find a good way of dynamically getting the android version letter
 	links := regexp.MustCompile("http.*("+device+"-p).*([0-9]{3}-).*(.zip)").FindAllString(body, -1)
 	factoryImage = links[len(links)-1]
 	_, err = url.ParseRequestURI(factoryImage)
@@ -321,35 +329,38 @@ func flashDevices() {
 		wg.Add(1)
 		go func(device string) {
 			defer wg.Done()
-			fmt.Println("Flashing stock firmware on device " + device + "...")
 			platformToolCommand := *fastboot
-			platformToolCommand.Args = append(platformToolCommand.Args, "-s", device, "--slot", "all", "flash", "bootloader", bootloader)
-			err := platformToolCommand.Run()
-			if err != nil {
-				fmt.Println("Failed to flash stock bootloader on device " + device)
-				return
-			}
-			platformToolCommand = *fastboot
-			platformToolCommand.Args = append(platformToolCommand.Args, "-s", device, "reboot-bootloader")
-			_ = platformToolCommand.Run()
-			time.Sleep(5 * time.Second)
-			platformToolCommand = *fastboot
-			platformToolCommand.Args = append(platformToolCommand.Args, "-s", device, "--slot", "all", "flash", "radio", radio)
-			err = platformToolCommand.Run()
-			if err != nil {
-				fmt.Println("Failed to flash stock radio on device " + device)
-				return
-			}
-			platformToolCommand = *fastboot
-			platformToolCommand.Args = append(platformToolCommand.Args, "-s", device, "reboot-bootloader")
-			_ = platformToolCommand.Run()
-			time.Sleep(5 * time.Second)
-			platformToolCommand = *fastboot
-			platformToolCommand.Args = append(platformToolCommand.Args, "-s", device, "--skip-reboot", "update", image)
-			err = platformToolCommand.Run()
-			if err != nil {
-				fmt.Println("Failed to flash stock image on device " + device)
-				return
+			err := errors.New("")
+			if factoryImage != "" {
+				fmt.Println("Flashing stock firmware on device " + device + "...")
+				platformToolCommand.Args = append(platformToolCommand.Args, "-s", device, "--slot", "all", "flash", "bootloader", bootloader)
+				err := platformToolCommand.Run()
+				if err != nil {
+					fmt.Println("Failed to flash stock bootloader on device " + device)
+					return
+				}
+				platformToolCommand = *fastboot
+				platformToolCommand.Args = append(platformToolCommand.Args, "-s", device, "reboot-bootloader")
+				_ = platformToolCommand.Run()
+				time.Sleep(5 * time.Second)
+				platformToolCommand = *fastboot
+				platformToolCommand.Args = append(platformToolCommand.Args, "-s", device, "--slot", "all", "flash", "radio", radio)
+				err = platformToolCommand.Run()
+				if err != nil {
+					fmt.Println("Failed to flash stock radio on device " + device)
+					return
+				}
+				platformToolCommand = *fastboot
+				platformToolCommand.Args = append(platformToolCommand.Args, "-s", device, "reboot-bootloader")
+				_ = platformToolCommand.Run()
+				time.Sleep(5 * time.Second)
+				platformToolCommand = *fastboot
+				platformToolCommand.Args = append(platformToolCommand.Args, "-s", device, "--skip-reboot", "update", image)
+				err = platformToolCommand.Run()
+				if err != nil {
+					fmt.Println("Failed to flash stock image on device " + device)
+					return
+				}
 			}
 			platformToolCommand = *fastboot
 			platformToolCommand.Args = append(platformToolCommand.Args, "-s", device, "reboot-bootloader")
@@ -375,33 +386,35 @@ func flashDevices() {
 		}(device)
 	}
 	wg.Wait()
-	for _, device := range devices {
-		fmt.Println("Locking device " + device + " bootloader...")
-		platformToolCommand := *fastboot
-		platformToolCommand.Args = append(platformToolCommand.Args, "-s", device, "erase", "avb_custom_key")
-		err := platformToolCommand.Run()
-		if err != nil {
-			fmt.Println("Failed to erase avb_custom_key. Exiting...")
-			return
+	if altosKey != "" {
+		for _, device := range devices {
+			fmt.Println("Locking device " + device + " bootloader...")
+			platformToolCommand := *fastboot
+			platformToolCommand.Args = append(platformToolCommand.Args, "-s", device, "erase", "avb_custom_key")
+			err := platformToolCommand.Run()
+			if err != nil {
+				fmt.Println("Failed to erase avb_custom_key. Exiting...")
+				return
+			}
+			platformToolCommand = *fastboot
+			platformToolCommand.Args = append(platformToolCommand.Args, "-s", device, "flash", "avb_custom_key", altosKey)
+			err = platformToolCommand.Run()
+			if err != nil {
+				fmt.Println("Failed to flash avb_custom_key. Exiting...")
+				return
+			}
+			fmt.Println("Please use the volume and power keys on the device to confirm.")
+			platformToolCommand = *fastboot
+			platformToolCommand.Args = append(platformToolCommand.Args, "-s", device, "flashing", "lock")
+			err = platformToolCommand.Run()
+			if err != nil {
+				fmt.Println("Failed to lock device bootloader. Exiting...")
+				return
+			}
+			time.Sleep(5 * time.Second)
+			fmt.Print("Press enter to continue")
+			_, _ = fmt.Scanln(&input)
 		}
-		platformToolCommand = *fastboot
-		platformToolCommand.Args = append(platformToolCommand.Args, "-s", device, "flash", "avb_custom_key", altosKey)
-		err = platformToolCommand.Run()
-		if err != nil {
-			fmt.Println("Failed to flash avb_custom_key. Exiting...")
-			return
-		}
-		fmt.Println("Please use the volume and power keys on the device to confirm.")
-		platformToolCommand = *fastboot
-		platformToolCommand.Args = append(platformToolCommand.Args, "-s", device, "flashing", "lock")
-		err = platformToolCommand.Run()
-		if err != nil {
-			fmt.Println("Failed to lock device bootloader. Exiting...")
-			return
-		}
-		time.Sleep(5 * time.Second)
-		fmt.Print("Press enter to continue")
-		_, _ = fmt.Scanln(&input)
 	}
 	for _, device := range devices {
 		fmt.Println("Rebooting " + device + "...")
